backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.3.0
bcrypt==4.0.1
